Static Site
This site was created without the use of JavaScript, as I did not learn it at the time. I had to create 4-12 pages to my site and they all had to follow the laws of sites. My favorite page is The Tower because I was able to make some creative choices with it.